Make Codeforces test-parsing more flexible
Older problems (such as 263A) separate testcase lines with
the <br> tag. Replace such tags with newline before parsing
the Html document. Newer problems should not be affected,
since they place the correctly formatted test data by itself
in the <pre> tag.

Older problems also don't start tests with a newline. Remove
these only if they exist. (we assume that the only reason for
a testcase to start with a newline is that it belongs to a new
problem).